# smart-hospital/api/openapi.yaml
openapi: 3.0.0
info:
  title: Smart Hospital Operations API
  description: Mock REST API for hospital doctor search, available slots, department inquiries, and appointment management.
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: Local Mock Hospital Backend Server

paths:
  /doctors:
    get:
      summary: Search and filter doctors
      operationId: search_doctors
      description: Returns a list of doctors filtered by department or query string.
      parameters:
        - name: department
          in: query
          required: false
          description: Department or specialty to filter by
          schema:
            type: string
        - name: query
          in: query
          required: false
          description: Search keyword matching doctor name, specialty, or bio
          schema:
            type: string
      responses:
        '200':
          description: Matching doctors list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Doctor'

  /doctors/{id}:
    get:
      summary: Get doctor details by ID
      operationId: get_doctor_by_id
      parameters:
        - name: id
          in: path
          required: true
          description: Unique doctor ID
          schema:
            type: string
      responses:
        '200':
          description: Doctor details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Doctor'

  /departments:
    get:
      summary: List all hospital departments
      operationId: get_departments
      description: Retrieves department names, descriptions, locations, and extensions.
      responses:
        '200':
          description: Department list

  /slots:
    get:
      summary: Get available appointment slots
      operationId: get_available_slots
      description: Returns open appointment time slots optionally filtered by doctor, department, or date.
      parameters:
        - name: doctor_id
          in: query
          required: false
          schema:
            type: string
        - name: department
          in: query
          required: false
          schema:
            type: string
        - name: date
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Available appointment slots list

  /appointments:
    get:
      summary: Get patient appointment history
      operationId: get_appointment_history
      parameters:
        - name: patient_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Patient appointment history
    post:
      summary: Book a new appointment
      operationId: book_appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patient_id, doctor_id, slot_id]
              properties:
                patient_id:
                  type: string
                doctor_id:
                  type: string
                slot_id:
                  type: string
                notes:
                  type: string
      responses:
        '201':
          description: Appointment successfully created

  /appointments/{id}:
    delete:
      summary: Cancel an existing appointment
      operationId: cancel_appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Appointment cancellation status confirmation

components:
  schemas:
    Doctor:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        department:
          type: string
        experience_years:
          type: integer
        rating:
          type: number
        bio:
          type: string